chore(applicationset): move webhook code to its own package
